Map Class
Adding python class to represent map as 0.25 inch squares and 3 inch squares.
When a 'map' quarter inch is determined to be occupied, we make the entire 3 inch block set to true as well. if we set an occupancy grid cell to true, all the 12X12 map points that correspond to this get set. TODO we might want to not update the map when occupancy grid is updated, only the other way areound